+++ lib/components/BestCourseTop3.dart
@@ -5,6 +5,7 @@ class BestCourseTop3 extends StatelessWidget {
   Widget build(BuildContext context) {
     final double screenWidth = MediaQuery.of(context).size.width;
     final double screenHeight = MediaQuery.of(context).size.height;
+    final PageController controller = PageController(viewportFraction: 0.9);
 
     return Column(
       crossAxisAlignment: CrossAxisAlignment.stretch,
@@ -20,8 +21,8 @@ class BestCourseTop3 extends StatelessWidget {
         SizedBox(height: screenHeight * 0.01),
         Container(
           height: screenHeight * 0.25,
-          child: ListView(
-            scrollDirection: Axis.horizontal,
+          child: PageView(
+            controller: controller,
             children: [
               RecommendedCourseCard(
                 imagePath: 'assets/course1.png',
@@ -42,7 +43,6 @@ class BestCourseTop3 extends StatelessWidget {
                 mbti: 'ESTJ',
               ),
               Container(
-                width: screenWidth * 0.3,
                 alignment: Alignment.center,
                 child: TextButton(
                   onPressed: () {},
@@ -93,11 +93,10 @@ class RecommendedCourseCard extends StatelessWidget {
     final double screenHeight = MediaQuery.of(context).size.height;
 
     return Container(
-      width: screenWidth * 0.9, //사진 가로크기
-      margin: EdgeInsets.only(
-          left: screenWidth * 0.05, right: screenWidth * 0.05), //양옆 공백
+      width: screenWidth * 0.9, // 사진 가로크기
+      margin: EdgeInsets.symmetric(horizontal: screenWidth * 0.05), // 양옆 공백
       decoration: BoxDecoration(
-        borderRadius: BorderRadius.circular(10), //사진 모따기 반경
+        borderRadius: BorderRadius.circular(10), // 사진 모따기 반경
         image: DecorationImage(
           image: AssetImage(imagePath),
           fit: BoxFit.cover,
@@ -106,13 +105,13 @@ class RecommendedCourseCard extends StatelessWidget {
       child: Stack(
         children: [
           Positioned(
-            bottom: screenHeight * 0.01, //사진설명텍스트 밑박스에서 거리
-            left: screenWidth * 0.025, //사진설명텍스트 박스기준왼쪽 거리
+            bottom: screenHeight * 0.01, // 사진설명텍스트 밑박스에서 거리
+            left: screenWidth * 0.025, // 사진설명텍스트 박스기준왼쪽 거리
             child: Text(
               '$mbti\n$title\n$location',
               style: TextStyle(
                 color: Colors.white,
-                fontSize: screenHeight * 0.02, //텍스트 크기
+                fontSize: screenHeight * 0.02, // 텍스트 크기
                 fontWeight: FontWeight.bold,
                 shadows: [
                   Shadow(
