+++ lib/main.dart
@@ -1,4 +1,31 @@
 import 'package:flutter/material.dart';
-import 'mbti_selection_page.dart'; // MBTI 선택 페이지 파일을 import
+import 'components/basic_frame_page.dart'; // 기본 프레임 페이지 파일을 import
 
 void main() => runApp(MBTISelectionApp());
+
+class MBTISelectionApp extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      theme: ThemeData(
+        fontFamily: 'Pretendard',
+        textTheme: TextTheme(
+          bodyLarge: TextStyle(fontWeight: FontWeight.bold),
+          bodyMedium: TextStyle(fontWeight: FontWeight.bold),
+          displayLarge: TextStyle(fontWeight: FontWeight.bold),
+          displayMedium: TextStyle(fontWeight: FontWeight.bold),
+          displaySmall: TextStyle(fontWeight: FontWeight.bold),
+          headlineMedium: TextStyle(fontWeight: FontWeight.bold),
+          headlineSmall: TextStyle(fontWeight: FontWeight.bold),
+          titleLarge: TextStyle(fontWeight: FontWeight.bold),
+          titleMedium: TextStyle(fontWeight: FontWeight.bold),
+          titleSmall: TextStyle(fontWeight: FontWeight.bold),
+          labelLarge: TextStyle(fontWeight: FontWeight.bold),
+          labelMedium: TextStyle(fontWeight: FontWeight.bold),
+          labelSmall: TextStyle(fontWeight: FontWeight.bold),
+        ),
+      ),
+      home: BasicFramePage(),
+    );
+  }
+}
