+++ lib/pages/course_generation_page.dart
