+++ lib/pages/mbti_selection_page.dart
@@ -1,35 +1,5 @@
 import 'package:flutter/material.dart';
 import 'calendar_selection_page.dart'; // 달력 선택 페이지 파일을 import
-import 'loginpage.dart'; // 로그인 페이지 파일을 import
-import 'BestCourseTop3.dart'; // 인기 추천코스 파일을 import
-import 'GyeongNamRecommend.dart'; // 경상남도 행사 & 축제 파일을 import
-
-class MBTISelectionApp extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      theme: ThemeData(
-        fontFamily: 'Pretendard',
-        textTheme: TextTheme(
-          bodyLarge: TextStyle(fontWeight: FontWeight.bold),
-          bodyMedium: TextStyle(fontWeight: FontWeight.bold),
-          displayLarge: TextStyle(fontWeight: FontWeight.bold),
-          displayMedium: TextStyle(fontWeight: FontWeight.bold),
-          displaySmall: TextStyle(fontWeight: FontWeight.bold),
-          headlineMedium: TextStyle(fontWeight: FontWeight.bold),
-          headlineSmall: TextStyle(fontWeight: FontWeight.bold),
-          titleLarge: TextStyle(fontWeight: FontWeight.bold),
-          titleMedium: TextStyle(fontWeight: FontWeight.bold),
-          titleSmall: TextStyle(fontWeight: FontWeight.bold),
-          labelLarge: TextStyle(fontWeight: FontWeight.bold),
-          labelMedium: TextStyle(fontWeight: FontWeight.bold),
-          labelSmall: TextStyle(fontWeight: FontWeight.bold),
-        ),
-      ),
-      home: MBTISelectionPage(),
-    );
-  }
-}
 
 class MBTISelectionPage extends StatefulWidget {
   @override
@@ -43,7 +13,6 @@ class _MBTISelectionPageState extends State<MBTISelectionPage> {
     'TF': null,
     'JP': null,
   };
-  int _selectedIndex = 0;
 
   void toggleSelection(String group, String mbti) {
     setState(() {
@@ -68,122 +37,59 @@ class _MBTISelectionPageState extends State<MBTISelectionPage> {
     }
   }
 
-  void _onItemTapped(int index) {
-    setState(() {
-      _selectedIndex = index;
-    });
-    if (index == 2) {
-      // '내 정보' 탭이 클릭되었을 때
-      Navigator.push(
-        context,
-        MaterialPageRoute(builder: (context) => LoginScreen()),
-      );
-    }
-  }
-
   @override
   Widget build(BuildContext context) {
     final double screenWidth = MediaQuery.of(context).size.width;
     final double screenHeight = MediaQuery.of(context).size.height;
 
-    return Scaffold(
-      backgroundColor: Colors.white,
-      appBar: PreferredSize(
-        preferredSize: Size.fromHeight(screenHeight * 0.08),
-        child: Container(
-          color: Colors.white,
-          child: AppBar(
-            centerTitle: true,
-            title: Image.asset('assets/mbtilogo.jpg',
-                height: screenHeight * 0.05), // MBTI 로고 경로
-            backgroundColor: Colors.white,
-            elevation: 0,
-            actions: [
-              IconButton(
-                icon: Image.asset('assets/alarm.jpg'), // 알림 아이콘 경로
-                onPressed: () {},
-              ),
-            ],
-          ),
+    return Column(
+      crossAxisAlignment: CrossAxisAlignment.stretch,
+      children: [
+        Divider(
+          color: Color(0xFFE4E4E4), // 실선 색상 설정
+          thickness: 1,
+          height: 1,
         ),
-      ),
-      body: SingleChildScrollView(
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.stretch,
+        SizedBox(height: screenHeight * 0.03),
+        Row(
           children: [
-            Divider(
-              color: Color(0xFFE4E4E4), // 실선 색상 설정
-              thickness: 1,
-              height: 1,
-            ),
-            SizedBox(height: screenHeight * 0.03),
-            Row(
-              children: [
-                SizedBox(width: screenWidth * 0.05),
-                Image.asset('assets/animation.png',
-                    height: screenHeight * 0.07), // 사람 아이콘 경로
-                SizedBox(width: screenWidth * 0.02),
-                Expanded(
-                  child: Container(
-                    padding: EdgeInsets.symmetric(
-                        vertical: screenHeight * 0.015,
-                        horizontal: screenWidth * 0.05),
-                    decoration: BoxDecoration(
-                      color: Colors.grey[200],
-                      borderRadius: BorderRadius.circular(20),
-                    ),
-                    child: Text(
-                      allGroupsSelected()
-                          ? '${selectedMBTI.values.join()} OOO님!\n오늘은 경상남도 어디로 떠나볼까요?'
-                          : '경상남도 추천지를 원하시면 MBTI를 선택해주세요 !!',
-                      style: TextStyle(fontSize: screenHeight * 0.02),
-                    ),
-                  ),
+            SizedBox(width: screenWidth * 0.05),
+            Image.asset('assets/animation.png',
+                height: screenHeight * 0.07), // 사람 아이콘 경로
+            SizedBox(width: screenWidth * 0.02),
+            Expanded(
+              child: Container(
+                padding: EdgeInsets.symmetric(
+                    vertical: screenHeight * 0.015,
+                    horizontal: screenWidth * 0.05),
+                decoration: BoxDecoration(
+                  color: Colors.grey[200],
+                  borderRadius: BorderRadius.circular(20),
                 ),
-                SizedBox(width: screenWidth * 0.05),
-              ],
+                child: Text(
+                  allGroupsSelected()
+                      ? '${selectedMBTI.values.join()} OOO님!\n오늘은 경상남도 어디로 떠나볼까요?'
+                      : '경상남도 추천지를 원하시면 MBTI를 선택해주세요 !!',
+                  style: TextStyle(fontSize: screenHeight * 0.02),
+                ),
+              ),
             ),
-            SizedBox(height: screenHeight * 0.02),
-            Text(
-              'MBTI 맞춤형 간단 추천 코스',
-              style: TextStyle(
-                  fontSize: screenHeight * 0.03, fontWeight: FontWeight.bold),
-              textAlign: TextAlign.center,
-            ),
-            StepOneWidget(
-              selectedMBTI: selectedMBTI,
-              toggleSelection: toggleSelection,
-              onNextPressed: navigateToCalendarPage,
-            ),
-            SizedBox(height: screenHeight * 0.02), // 추가 여백
-            BestCourseTop3(), // 인기 추천코스 BEST 3 위젯 사용
-            SizedBox(height: screenHeight * 0.02), // 추가 여백
-            GyeongNamRecommend(), // 경상남도 행사 & 축제 위젯 사용
+            SizedBox(width: screenWidth * 0.05),
           ],
         ),
-      ),
-      bottomNavigationBar: BottomNavigationBar(
-        backgroundColor: Colors.white, // 하단바 배경색을 흰색으로 설정
-        items: [
-          BottomNavigationBarItem(
-            icon: Image.asset('assets/home.jpg',
-                height: screenHeight * 0.04), // 홈 아이콘 경로
-            label: '홈',
-          ),
-          BottomNavigationBarItem(
-            icon: Image.asset('assets/location1.png',
-                height: screenHeight * 0.04), // 여행 일정 아이콘 경로
-            label: '여행 일정',
-          ),
-          BottomNavigationBarItem(
-            icon: Image.asset('assets/myprofile.png',
-                height: screenHeight * 0.04), // 내 정보 아이콘 경로
-            label: '내 정보',
-          ),
-        ],
-        currentIndex: _selectedIndex,
-        onTap: _onItemTapped,
-      ),
+        SizedBox(height: screenHeight * 0.02),
+        Text(
+          'MBTI 맞춤형 간단 추천 코스',
+          style: TextStyle(
+              fontSize: screenHeight * 0.03, fontWeight: FontWeight.bold),
+          textAlign: TextAlign.center,
+        ),
+        StepOneWidget(
+          selectedMBTI: selectedMBTI,
+          toggleSelection: toggleSelection,
+          onNextPressed: navigateToCalendarPage,
+        ),
+      ],
     );
   }
 }
